+++ src/main/java/com/example/gruppe4bilabonnement/models/DamageReport.java
@@ -4,26 +4,26 @@ public class DamageReport {
     public DamageReport() {
     }
 
-    public DamageReport(Long id, Long carId, String damageText, Double price) {
+    public DamageReport(int id, int carId, String damageText, Double price) {
         this.id = id;
         this.carId = carId;
         this.damageText = damageText;
         this.price = price;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
-    public Long getCarId() {
+    public int getCarId() {
         return carId;
     }
 
-    public void setCarId(Long carId) {
+    public void setCarId(int carId) {
         this.carId = carId;
     }
 
@@ -43,8 +43,8 @@ public class DamageReport {
         this.price = price;
     }
 
-    private Long id;
-    private Long carId;
+    private int id;
+    private int carId;
     private String damageText;
     private Double price;
